FixBug

+++ W10_resetD_dcgan/dcgan.py
@@ -352,13 +352,14 @@ for j, epoch in enumerate(range(start_epoch,opt.n_epochs+1)):
 			save_model(discriminator,optimizer_D,epoch,opt.model_save_path+"/tmp_D.pt")
 			trainG = True
 			nb_save = nb_save+1
-	elif D_x[j] > 0.7 and D_G_z[j] < 0.2: 
+	if D_x[j] > 0.7 and D_G_z[j] < 0.2: 
 	#elif D_x[j] < 0.2 and D_G_z[j] > 0.6: 
-		print("Load previous model for reset D")
-		load_model(discriminator,optimizer_D,opt.model_save_path+"/tmp_D.pt")
-		os.remove(opt.model_save_path+"/tmp_D.pt")
-		trainG = False
-		nb_load = nb_load+1
+		if os.path.exists(opt.model_save_path+"/tmp_D.pt"):
+			print("Load previous model for reset D")
+			load_model(discriminator,optimizer_D,opt.model_save_path+"/tmp_D.pt")
+			os.remove(opt.model_save_path+"/tmp_D.pt")
+			trainG = False
+			nb_load = nb_load+1
 	
 	# Save samples
 	if epoch % opt.sample_interval == 0:
